refactor: 클래스 소스 파일 분리
- Application에서 클래스 별도 파일로 분리
- 컨벤션 고려해서 depth, import 수정
